# HealthCareApiServer/WebApp/client-app/openapi.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "My API",
    "version": "v1"
  },
  "servers": [
    {
      "url": "https://localhost:44351"
    }
  ],
  "paths": {
...
    "/api/PrescriptionIntake/ByPrescription/{prescriptionId}": {
      "get": {
        "tags": [
          "PrescriptionIntake"
        ],
        "operationId": "PrescriptionIntake_GetByPrescription",
        "parameters": [
          {
            "name": "prescriptionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PrescriptionIntakeModel"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/PrescriptionIntake/Mark": {
      "post": {
        "tags": [
          "PrescriptionIntake"
        ],
        "operationId": "PrescriptionIntake_Mark",
        "requestBody": {
          "x-name": "model",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PrescriptionIntakeModel"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrescriptionIntakeModel"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "PrescriptionIntakeModel": {
        "allOf": [
          {
            "$ref": "#/components/schemas/EntityModel"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "required": [
              "prescriptionId"
            ],
            "properties": {
              "prescriptionId": {
                "type": "integer",
                "format": "int32"
              },
              "scheduledFor": {
                "type": "string",
                "format": "date-time"
              },
              "takenAt": {
                "type": "string",
                "format": "date-time",
                "nullable": true
              },
              "prescriptionIsActive": {
                "type": "boolean"
              }
            }
          }
        ]
      },
      "EntityModel": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "createdOn": {
            "type": "string",
            "format": "date-time"
          },
          "createdBy": {
            "type": "string",
            "nullable": true
          },
          "modifiedOn": {
            "type": "string",
            "format": "date-time"
          },
          "modifiedBy": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
    },
    "securitySchemes": {
      "JWT": {
        "type": "http",
        "description": "Type into the textbox: Bearer {your JWT token}.",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "JWT": []
    }
  ]
}