# HealthCareApiServer/WebApp/client-app/src/app/api/models/prescription-intake-model.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { EntityModel } from '../models/entity-model';
export type PrescriptionIntakeModel = EntityModel & {
'prescriptionId': number;
'scheduledFor'?: string;
'takenAt'?: string | null;
'prescriptionIsActive'?: boolean;
};